Renamed modal directory to upper case
